+++ src/pycentauri/__init__.py
@@ -25,4 +25,4 @@ __all__ = [
     "discover",
 ]
 
-__version__ = "0.4.1"
+__version__ = "0.4.2"